refactored services to align with database schema
