+++ server/config.py
@@ -35,6 +35,8 @@ class Config:
     NEO4J_URI = os.environ.get('NEO4J_URI')
     NEO4J_USERNAME = os.environ.get('NEO4J_USERNAME')
     NEO4J_PASSWORD = os.environ.get('NEO4J_PASSWORD')
+    NEO4J_DATABASE = os.environ.get('NEO4J_DATABASE', 'neo4j')
+
     
     # Ensure data directory exists
     os.makedirs(DATA_DIR, exist_ok=True)
