+++ server/database/__init__.py
@@ -1,0 +1,3 @@
+from .db_connection import db_connection  # ✅ Match your actual filename
+
+__all__ = ['db_connection']
