+++ server/services/conversation_service.py
@@ -1,24 +1,14 @@
-import openai
-from flask import current_app
-from utils.conversation_utils import (
+
+from ..utils.conversation_utils import (
     load_user_conversations, save_user_conversations, add_message_to_conversation,
     get_recent_messages, should_summarize_conversation, get_current_conversation,
     start_new_conversation
 )
+from .llm_service import LLMService
 
 class ConversationService:
     def __init__(self):
-        self._openai_client = None
-    
-    @property
-    def openai_client(self):
-        """Lazy initialization of OpenAI client"""
-        if self._openai_client is None:
-            api_key = current_app.config.get('OPENAI_API_KEY')
-            if not api_key:
-                raise ValueError("OPENAI_API_KEY not configured")
-            self._openai_client = openai.OpenAI(api_key=api_key)
-        return self._openai_client
+        self.llm_service = LLMService()
     
     def generate_conversation_summary(self, messages):
         """Generate a summary of conversation messages using GPT-4"""
